Added stream user data. Chfnge UI Delete interval timer

+++ BinanceAPI/Model/DataBinance.cs
@@ -8,7 +8,7 @@ namespace BinanceAPI.Model
     {
 
         public string symbol { get; set; }
-        public double lastPrice { get; set;}
+        public decimal lastPrice { get; set;}
         //public string lastPrice { get; set; }
         //public string highPrice { get; set; }
         // public string lowPrice { get; set; }
